made updates to the fasting app functions AllowAny

+++ fitness_app/measurements/models.py
@@ -3,6 +3,6 @@ from django.db import models
 class Measurements(models.Model):
     user = models.IntegerField
     body_part = models.CharField(max_length=30)
-    measurement = models.DecimalField(max_digits=6, decimal_places=1)
+    measurement = models.IntegerField()
     date = models.DateField()
